Separate out job interface implementation to serializer (#1)
* separate out job interface implementation to serializer

* Remove accidentally added binary

* Some renaming

---------
